+++ src/app/(with-side)/[type]/[id]/page.tsx
@@ -4,9 +4,9 @@ import { useEffect, useRef, useState } from 'react';
 import { useSearchParams, useParams, useRouter } from 'next/navigation';
 
 import ComponentFrame from '@/components/common/componentFrame';
-import RecommendedContent from '@/components/common/recommendedContent';
+import RecommendedContent from '@/components/common/recommended-content-temp';
 import DetailContent from '@/components/features/detail/content/DetailContent';
-import CommentSection from '@/components/features/detail/comment/comment-section-temp';
+import CommentSection from '@/components/features/detail/comment/comment-section';
 import FetchingFailedCard from '@/components/common/fetching-failed-card';
 import PaginationBar from '@/components/common/pagination-bar.tsx/pagination-bar';
