+++ src/components/common/category-tab/news-item.tsx
@@ -2,7 +2,7 @@ import { getTimeAgo } from '@/lib/utils/getTimeAgo';
 import { NewsItemDto } from '@/services/apis/news/dto';
 import Image from 'next/image';
 import Link from 'next/link';
-import SanitizedContent from '../sanitized-content-temp';
+import SanitizedContent from '../sanitized-content';
 
 export default function NewsItem({
 	pk,
