+++ src/components/common/floating-writing-button.tsx
