+++ src/components/common/sanitized-content.tsx
